change buffer lengths of sysctl(2) to be size_t's
SCCS-vsn: 7.30

+++ sys/kern/kern_clock.c
@@ -30,7 +30,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- *	@(#)kern_clock.c	7.29 (Berkeley) 02/04/93
+ *	@(#)kern_clock.c	7.30 (Berkeley) 03/04/93
  */
 
 #include <sys/param.h>
@@ -486,7 +486,7 @@ statclock(frame)
  */
 sysctl_clockrate(where, sizep)
 	register char *where;
-	int *sizep;
+	size_t *sizep;
 {
 	struct clockinfo clkinfo;
